Move test setup and utils to test dir, modify tsconfig and build process to work with new structure
