+++ .gitignore
@@ -1,4 +1,5 @@
 # Built files
+build
 generators
 
 # Environment Variables -- May Contain Secrets
