+++ jest.config.ts
@@ -3,7 +3,7 @@ import type { Config } from "jest";
 const config: Config = {
   coverageDirectory: "<rootDir>/reports/coverage",
   preset: "ts-jest",
-  setupFilesAfterEnv: ["./src/__tests__/__setup__/jest.setup.ts"],
+  setupFilesAfterEnv: ["<rootDir>/test/setup/jest.setup.ts"],
   testEnvironment: "node",
   testMatch: ["**/?(*.)+(spec|test).[jt]s?(x)"],
 };
