+++ test/matchers/yeoman/toHaveDevDependencyWithComment.matcher.ts
