+++ test/utils/testGenerators/DevDependenciesTestGenerator.ts
@@ -1,4 +1,4 @@
-import { BaseGenerator } from "../../../shared/index";
+import { BaseGenerator } from "../../../src/shared";
 
 export class DevDependenciesTestGenerator extends BaseGenerator {
   public async writing() {
