+++ test/utils/testGenerators/EmptyTestGenerator.ts
@@ -1,4 +1,4 @@
-import { BaseGenerator } from "../../../shared/index";
+import { BaseGenerator } from "../../../src/shared";
 
 export class EmptyTestGenerator extends BaseGenerator {
   public writing() {
